[C++] Extract the internals of the RunEndEncode kernel out of compute/
