+++ cpp/src/arrow/CMakeLists.txt
@@ -772,7 +772,6 @@ if(ARROW_COMPUTE)
        compute/kernels/hash_aggregate_numeric.cc
        compute/kernels/hash_aggregate_pivot.cc
        compute/kernels/pivot_internal.cc
-       compute/kernels/ree_util_internal.cc
        compute/kernels/scalar_arithmetic.cc
        compute/kernels/scalar_boolean.cc
        compute/kernels/scalar_compare.cc
