+++ cpp/src/arrow/c/CMakeLists.txt
@@ -15,28 +15,7 @@
 # specific language governing permissions and limitations
 # under the License.
 
-# TODO(GH-37221): Remove compute dependency for REE requirements on bridge_test
-set(ARROW_TEST_LINK_LIBS "")
-
-if(ARROW_TEST_LINKAGE STREQUAL "static")
-  list(APPEND ARROW_TEST_LINK_LIBS ${ARROW_TEST_STATIC_LINK_LIBS})
-else()
-  list(APPEND ARROW_TEST_LINK_LIBS ${ARROW_TEST_SHARED_LINK_LIBS})
-endif()
-
-if(ARROW_COMPUTE)
-  if(ARROW_TEST_LINKAGE STREQUAL "static")
-    list(APPEND ARROW_TEST_LINK_LIBS arrow_compute_static arrow_compute_testing)
-  else()
-    list(APPEND ARROW_TEST_LINK_LIBS arrow_compute_shared arrow_compute_testing)
-  endif()
-endif()
-
-add_arrow_test(bridge_test
-               PREFIX
-               "arrow-c"
-               STATIC_LINK_LIBS
-               ${ARROW_TEST_LINK_LIBS})
+add_arrow_test(bridge_test PREFIX "arrow-c")
 add_arrow_test(dlpack_test)
 
 add_arrow_benchmark(bridge_benchmark)
