+++ cpp/src/arrow/c/bridge_test.cc
@@ -31,6 +31,7 @@
 #include "arrow/c/bridge.h"
 #include "arrow/c/helpers.h"
 #include "arrow/c/util_internal.h"
+#include "arrow/json/from_string.h"
 #include "arrow/memory_pool.h"
 #include "arrow/testing/builder.h"
 #include "arrow/testing/extension_type.h"
@@ -45,12 +46,8 @@
 #include "arrow/util/logging_internal.h"
 #include "arrow/util/macros.h"
 #include "arrow/util/range.h"
+#include "arrow/util/ree_util.h"
 #include "arrow/util/thread_pool.h"
-
-// TODO(GH-37221): Remove these ifdef checks when compute dependency is removed
-#ifdef ARROW_COMPUTE
-#  include "arrow/compute/api_vector.h"
-#endif
 
 namespace arrow {
 
@@ -472,7 +469,6 @@ TEST_F(TestSchemaExport, Union) {
              {ARROW_FLAG_NULLABLE});
 }
 
-#ifdef ARROW_COMPUTE
 TEST_F(TestSchemaExport, RunEndEncoded) {
   TestNested(run_end_encoded(int16(), uint8()), {"+r", "s", "C"},
              {"", "run_ends", "values"}, {ARROW_FLAG_NULLABLE, 0, ARROW_FLAG_NULLABLE});
@@ -484,7 +480,6 @@ TEST_F(TestSchemaExport, RunEndEncoded) {
              {"", "run_ends", "values", "item"},
              {ARROW_FLAG_NULLABLE, 0, ARROW_FLAG_NULLABLE, ARROW_FLAG_NULLABLE});
 }
-#endif
 
 std::string GetIndexFormat(Type::type type_id) {
   switch (type_id) {
@@ -1080,21 +1075,10 @@ TEST_F(TestArrayExport, Union) {
   TestNested(type, data);
 }
 
-#ifdef ARROW_COMPUTE
-Result<std::shared_ptr<Array>> REEFromJSON(const std::shared_ptr<DataType>& ree_type,
-                                           const std::string& json) {
-  auto ree_type_ptr = checked_cast<const RunEndEncodedType*>(ree_type.get());
-  auto array = ArrayFromJSON(ree_type_ptr->value_type(), json);
-  ARROW_ASSIGN_OR_RAISE(
-      auto datum,
-      RunEndEncode(array, compute::RunEndEncodeOptions{ree_type_ptr->run_end_type()}));
-  return datum.make_array();
-}
-
 TEST_F(TestArrayExport, RunEndEncoded) {
   auto factory = []() {
-    return REEFromJSON(run_end_encoded(int32(), int8()),
-                       "[1, 2, 2, 3, null, null, null, 4]");
+    return ArrayFromJSON(run_end_encoded(int32(), int8()),
+                         "[1, 2, 2, 3, null, null, null, 4]");
   };
   TestNested(factory);
 }
@@ -1102,13 +1086,12 @@ TEST_F(TestArrayExport, RunEndEncoded) {
 TEST_F(TestArrayExport, RunEndEncodedSliced) {
   auto factory = []() -> Result<std::shared_ptr<Array>> {
     ARROW_ASSIGN_OR_RAISE(auto ree_array,
-                          REEFromJSON(run_end_encoded(int32(), int8()),
-                                      "[1, 2, 2, 3, null, null, null, 4]"));
+                          json::ArrayFromJSONString(run_end_encoded(int32(), int8()),
+                                                    "[1, 2, 2, 3, null, null, null, 4]"));
     return ree_array->Slice(1, 5);
   };
   TestNested(factory);
 }
-#endif
 
 TEST_F(TestArrayExport, Dictionary) {
   {
@@ -1432,16 +1415,14 @@ class TestDeviceArrayExport : public ::testing::Test {
     return [=]() { return ToDevice(mm, *ArrayFromJSON(type, json)->data()); };
   }
 
-#ifdef ARROW_COMPUTE
   static std::function<Result<std::shared_ptr<Array>>()> JSONREEArrayFactory(
       const std::shared_ptr<MemoryManager>& mm, std::shared_ptr<DataType> type,
       const char* json) {
     return [=]() -> Result<std::shared_ptr<Array>> {
-      ARROW_ASSIGN_OR_RAISE(auto result, REEFromJSON(type, json));
+      ARROW_ASSIGN_OR_RAISE(auto result, json::ArrayFromJSONString(type, json));
       return ToDevice(mm, *result->data());
     };
   }
-#endif
 
   template <typename ArrayFactory, typename ExportCheckFunc>
   void TestWithArrayFactory(ArrayFactory&& factory, ExportCheckFunc&& check_func) {
@@ -1680,7 +1661,6 @@ TEST_F(TestDeviceArrayExport, Union) {
   TestNested(mm, type, data);
 }
 
-#ifdef ARROW_COMPUTE
 TEST_F(TestDeviceArrayExport, RunEndEncoded) {
   std::shared_ptr<Device> device = std::make_shared<MyDevice>(1);
   auto mm = device->default_memory_manager();
@@ -1689,7 +1669,6 @@ TEST_F(TestDeviceArrayExport, RunEndEncoded) {
   const char* data = "[1, null, 2, 2, 4, 5]";
   TestNested(JSONREEArrayFactory(mm, type, data));
 }
-#endif
 
 TEST_F(TestDeviceArrayExport, Extension) {
   std::shared_ptr<Device> device = std::make_shared<MyDevice>(1);
@@ -2186,14 +2165,12 @@ TEST_F(TestSchemaImport, Map) {
   CheckImport(expected);
 }
 
-#ifdef ARROW_COMPUTE
 TEST_F(TestSchemaImport, RunEndEncoded) {
   FillPrimitive(AddChild(), "s", "run_ends");
   FillPrimitive(AddChild(), "I", "values");
   FillRunEndEncoded("+r");
   CheckImport(run_end_encoded(int16(), uint32()));
 }
-#endif
 
 TEST_F(TestSchemaImport, Dictionary) {
   FillPrimitive(AddChild(), "u");
@@ -3175,14 +3152,14 @@ TEST_F(TestArrayImport, Struct) {
   CheckImport(expected);
 }
 
-#ifdef ARROW_COMPUTE
 TEST_F(TestArrayImport, RunEndEncoded) {
   FillPrimitive(AddChild(), 5, 0, 0, run_ends_buffers5);
   FillPrimitive(AddChild(), 5, 0, 0, primitive_buffers_no_nulls5);
   FillRunEndEncoded(9, 0);
-  ASSERT_OK_AND_ASSIGN(auto expected,
-                       REEFromJSON(run_end_encoded(int16(), float32()),
-                                   "[0.0, 1.5, -2.0, -2.0, 3.0, 3.0, 3.0, 4.0, 4.0]"));
+  ASSERT_OK_AND_ASSIGN(
+      auto expected,
+      json::ArrayFromJSONString(run_end_encoded(int16(), float32()),
+                                "[0.0, 1.5, -2.0, -2.0, 3.0, 3.0, 3.0, 4.0, 4.0]"));
   ASSERT_OK(expected->ValidateFull());
   CheckImport(expected);
 }
@@ -3193,32 +3170,35 @@ TEST_F(TestArrayImport, RunEndEncodedWithOffset) {
   FillPrimitive(AddChild(), 3, 0, 2, run_ends_buffers5);
   FillPrimitive(AddChild(), 3, 0, 2, primitive_buffers_no_nulls5);
   FillRunEndEncoded(7, 0);
-  ASSERT_OK_AND_ASSIGN(auto expected,
-                       REEFromJSON(ree_type, "[-2.0, -2.0, -2.0, -2.0, 3.0, 3.0, 3.0]"));
+  ASSERT_OK_AND_ASSIGN(
+      auto expected,
+      json::ArrayFromJSONString(ree_type, "[-2.0, -2.0, -2.0, -2.0, 3.0, 3.0, 3.0]"));
   CheckImport(expected);
 
   // Offset in parent
   FillPrimitive(AddChild(), 5, 0, 0, run_ends_buffers5);
   FillPrimitive(AddChild(), 5, 0, 0, primitive_buffers_no_nulls5);
   FillRunEndEncoded(5, 2);
-  ASSERT_OK_AND_ASSIGN(expected, REEFromJSON(ree_type, "[-2.0, -2.0, 3.0, 3.0, 3.0]"));
+  ASSERT_OK_AND_ASSIGN(
+      expected, json::ArrayFromJSONString(ree_type, "[-2.0, -2.0, 3.0, 3.0, 3.0]"));
   CheckImport(expected);
 
   // Length in parent that cuts last run
   FillPrimitive(AddChild(), 5, 0, 0, run_ends_buffers5);
   FillPrimitive(AddChild(), 5, 0, 0, primitive_buffers_no_nulls5);
   FillRunEndEncoded(4, 2);
-  ASSERT_OK_AND_ASSIGN(expected, REEFromJSON(ree_type, "[-2.0, -2.0, 3.0, 3.0]"));
+  ASSERT_OK_AND_ASSIGN(expected,
+                       json::ArrayFromJSONString(ree_type, "[-2.0, -2.0, 3.0, 3.0]"));
   CheckImport(expected);
 
   // Offset in both children and parent
   FillPrimitive(AddChild(), 3, 0, 2, run_ends_buffers5);
   FillPrimitive(AddChild(), 3, 0, 2, primitive_buffers_no_nulls5);
   FillRunEndEncoded(4, 2);
-  ASSERT_OK_AND_ASSIGN(expected, REEFromJSON(ree_type, "[-2.0, -2.0, 3.0, 3.0]"));
+  ASSERT_OK_AND_ASSIGN(expected,
+                       json::ArrayFromJSONString(ree_type, "[-2.0, -2.0, 3.0, 3.0]"));
   CheckImport(expected);
 }
-#endif
 
 TEST_F(TestArrayImport, SparseUnion) {
   auto type = sparse_union({field("strs", utf8()), field("ints", int8())}, {43, 42});
@@ -3749,12 +3729,10 @@ TEST_F(TestSchemaRoundtrip, Union) {
   TestWithTypeFactory([&]() { return dense_union({f1, f2}, type_codes); });
 }
 
-#ifdef ARROW_COMPUTE
 TEST_F(TestSchemaRoundtrip, RunEndEncoded) {
   TestWithTypeFactory([]() { return run_end_encoded(int16(), float32()); });
   TestWithTypeFactory([]() { return run_end_encoded(int32(), list(float32())); });
 }
-#endif
 
 TEST_F(TestSchemaRoundtrip, Dictionary) {
   for (auto index_ty : all_dictionary_index_types()) {
@@ -4104,13 +4082,12 @@ TEST_F(TestArrayRoundtrip, Union) {
   }
 }
 
-#ifdef ARROW_COMPUTE
 TEST_F(TestArrayRoundtrip, RunEndEncoded) {
   {
     auto factory = []() -> Result<std::shared_ptr<Array>> {
-      ARROW_ASSIGN_OR_RAISE(auto ree_array,
-                            REEFromJSON(run_end_encoded(int32(), int8()),
-                                        "[1, 2, 2, 3, null, null, null, 4]"));
+      ARROW_ASSIGN_OR_RAISE(
+          auto ree_array, json::ArrayFromJSONString(run_end_encoded(int32(), int8()),
+                                                    "[1, 2, 2, 3, null, null, null, 4]"));
       return ree_array->Slice(1, 5);
     };
     TestWithArrayFactory(factory);
@@ -4130,7 +4107,6 @@ TEST_F(TestArrayRoundtrip, RunEndEncoded) {
     TestWithArrayFactory(factory);
   }
 }
-#endif
 
 TEST_F(TestArrayRoundtrip, Dictionary) {
   {
