+++ cpp/src/arrow/util/ree_util.h
@@ -17,14 +17,20 @@
 
 #pragma once
 
-#include <algorithm>
 #include <cassert>
 #include <cstdint>
+#include <memory>
+#include <tuple>
+#include <utility>
 
 #include "arrow/array/data.h"
+#include "arrow/result.h"
+#include "arrow/status.h"
+#include "arrow/type.h"
+#include "arrow/type_fwd.h"
 #include "arrow/type_traits.h"
-#include "arrow/util/checked_cast.h"
 #include "arrow/util/macros.h"
+#include "arrow/util/visibility.h"
 
 namespace arrow {
 namespace ree_util {
@@ -48,6 +54,7 @@ const RunEndCType* RunEnds(const ArraySpan& span) {
 /// All the checks complete in O(1) time. Consequently, this function:
 /// - DOES NOT check that run_ends is sorted and all-positive
 /// - DOES NOT check the actual contents of the run_ends and values arrays
+ARROW_EXPORT
 Status ValidateRunEndEncodedChildren(const RunEndEncodedType& type,
                                      int64_t logical_length,
                                      const std::shared_ptr<ArrayData>& run_ends_data,
@@ -55,155 +62,8 @@ Status ValidateRunEndEncodedChildren(const RunEndEncodedType& type,
                                      int64_t null_count, int64_t logical_offset);
 
 /// \brief Compute the logical null count of an REE array
+ARROW_EXPORT
 int64_t LogicalNullCount(const ArraySpan& span);
-
-namespace internal {
-
-/// \brief Uses binary-search to find the physical offset given a logical offset
-/// and run-end values
-///
-/// \return the physical offset or run_ends_size if the physical offset is not
-/// found in run_ends
-template <typename RunEndCType>
-int64_t FindPhysicalIndex(const RunEndCType* run_ends, int64_t run_ends_size, int64_t i,
-                          int64_t absolute_offset) {
-  assert(absolute_offset + i >= 0);
-  auto it = std::upper_bound(run_ends, run_ends + run_ends_size, absolute_offset + i);
-  int64_t result = std::distance(run_ends, it);
-  assert(result <= run_ends_size);
-  return result;
-}
-
-/// \brief Uses binary-search to calculate the range of physical values (and
-/// run-ends) necessary to represent the logical range of values from
-/// offset to length
-///
-/// \return a pair of physical offset and physical length
-template <typename RunEndCType>
-std::pair<int64_t, int64_t> FindPhysicalRange(const RunEndCType* run_ends,
-                                              int64_t run_ends_size, int64_t length,
-                                              int64_t offset) {
-  const int64_t physical_offset =
-      FindPhysicalIndex<RunEndCType>(run_ends, run_ends_size, 0, offset);
-  // The physical length is calculated by finding the offset of the last element
-  // and adding 1 to it, so first we ensure there is at least one element.
-  if (length == 0) {
-    return {physical_offset, 0};
-  }
-  const int64_t physical_index_of_last = FindPhysicalIndex<RunEndCType>(
-      run_ends + physical_offset, run_ends_size - physical_offset, length - 1, offset);
-
-  assert(physical_index_of_last < run_ends_size - physical_offset);
-  return {physical_offset, physical_index_of_last + 1};
-}
-
-/// \brief Uses binary-search to calculate the number of physical values (and
-/// run-ends) necessary to represent the logical range of values from
-/// offset to length
-template <typename RunEndCType>
-int64_t FindPhysicalLength(const RunEndCType* run_ends, int64_t run_ends_size,
-                           int64_t length, int64_t offset) {
-  auto [_, physical_length] =
-      FindPhysicalRange<RunEndCType>(run_ends, run_ends_size, length, offset);
-  // GH-37107: This is a workaround for GCC 7. GCC 7 doesn't ignore
-  // variables in structured binding automatically from unused
-  // variables when one of these variables are used.
-  // See also: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=81767
-  ARROW_UNUSED(_);
-  return physical_length;
-}
-
-/// \brief Find the physical index into the values array of the REE ArraySpan
-///
-/// This function uses binary-search, so it has a O(log N) cost.
-template <typename RunEndCType>
-int64_t FindPhysicalIndex(const ArraySpan& span, int64_t i, int64_t absolute_offset) {
-  const int64_t run_ends_size = RunEndsArray(span).length;
-  return FindPhysicalIndex(RunEnds<RunEndCType>(span), run_ends_size, i, absolute_offset);
-}
-
-/// \brief Find the physical length of an REE ArraySpan
-///
-/// The physical length of an REE is the number of physical values (and
-/// run-ends) necessary to represent the logical range of values from
-/// offset to length.
-///
-/// Avoid calling this function if the physical length can be established in
-/// some other way (e.g. when iterating over the runs sequentially until the
-/// end). This function uses binary-search, so it has a O(log N) cost.
-template <typename RunEndCType>
-int64_t FindPhysicalLength(const ArraySpan& span) {
-  return FindPhysicalLength(
-      /*run_ends=*/RunEnds<RunEndCType>(span),
-      /*run_ends_size=*/RunEndsArray(span).length,
-      /*length=*/span.length,
-      /*offset=*/span.offset);
-}
-
-template <typename RunEndCType>
-struct PhysicalIndexFinder;
-
-// non-inline implementations for each run-end type
-ARROW_EXPORT int64_t FindPhysicalIndexImpl16(PhysicalIndexFinder<int16_t>& self,
-                                             int64_t i);
-ARROW_EXPORT int64_t FindPhysicalIndexImpl32(PhysicalIndexFinder<int32_t>& self,
-                                             int64_t i);
-ARROW_EXPORT int64_t FindPhysicalIndexImpl64(PhysicalIndexFinder<int64_t>& self,
-                                             int64_t i);
-
-/// \brief Stateful version of FindPhysicalIndex() that caches the result of
-/// the previous search and uses it to optimize the next search.
-///
-/// When new queries for the physical index of a logical index come in,
-/// binary search is performed again but the first candidate checked is the
-/// result of the previous search (cached physical index) instead of the
-/// midpoint of the run-ends array.
-///
-/// If that test fails, internal::FindPhysicalIndex() is called with one of the
-/// partitions defined by the cached index. If the queried logical indices
-/// follow an increasing or decreasing pattern, this first test is much more
-/// effective in (1) finding the answer right away (close logical indices belong
-/// to the same runs) or (2) discarding many more candidates than probing
-/// the midpoint would.
-///
-/// The most adversarial case (i.e. alternating between 0 and length-1 queries)
-/// only adds one extra binary search probe when compared to always starting
-/// binary search from the midpoint without any of these optimizations.
-///
-/// \tparam RunEndCType The numeric type of the run-ends array.
-template <typename RunEndCType>
-struct PhysicalIndexFinder {
-  const ArraySpan array_span;
-  const RunEndCType* run_ends;
-  int64_t last_physical_index = 0;
-
-  explicit PhysicalIndexFinder(const ArrayData& data)
-      : array_span(data),
-        run_ends(RunEndsArray(array_span).template GetValues<RunEndCType>(1)) {
-    assert(CTypeTraits<RunEndCType>::ArrowType::type_id ==
-           ::arrow::internal::checked_cast<const RunEndEncodedType&>(*data.type)
-               .run_end_type()
-               ->id());
-  }
-
-  /// \brief Find the physical index into the values array of the REE array.
-  ///
-  /// \pre 0 <= i < array_span.length()
-  /// \param i the logical index into the REE array
-  /// \return the physical index into the values array
-  int64_t FindPhysicalIndex(int64_t i) {
-    if constexpr (std::is_same_v<RunEndCType, int16_t>) {
-      return FindPhysicalIndexImpl16(*this, i);
-    } else if constexpr (std::is_same_v<RunEndCType, int32_t>) {
-      return FindPhysicalIndexImpl32(*this, i);
-    } else {
-      static_assert(std::is_same_v<RunEndCType, int64_t>, "Unsupported RunEndCType.");
-      return FindPhysicalIndexImpl64(*this, i);
-    }
-  }
-};
-
-}  // namespace internal
 
 /// \brief Find the physical index into the values array of the REE ArraySpan
 ///
@@ -230,7 +90,36 @@ ARROW_EXPORT std::pair<int64_t, int64_t> FindPhysicalRange(const ArraySpan& span
                                                            int64_t offset,
                                                            int64_t length);
 
-// Publish PhysicalIndexFinder outside of the internal namespace.
+/// \brief Run-end encode an array
+///
+/// This function encodes an array into run-end encoded format. It uses an optimized
+/// two-pass algorithm: first counting runs, then preallocating buffers and writing.
+///
+/// \param[in] input The array to encode as ArrayData
+/// \param[in] run_end_type The integer type for run ends (int16, int32, or int64)
+/// \param[in] pool Memory pool for allocations
+/// \return A RunEndEncodedArrayData
+ARROW_EXPORT
+Result<std::shared_ptr<ArrayData>> RunEndEncodeArray(
+    const ArraySpan& input, const std::shared_ptr<DataType>& run_end_type,
+    MemoryPool* pool);
+
+/// \brief Run-end encode an array (convenience overload accepting ArrayData)
+ARROW_EXPORT
+Result<std::shared_ptr<ArrayData>> RunEndEncodeArray(
+    const std::shared_ptr<ArrayData>& input,
+    const std::shared_ptr<DataType>& run_end_type, MemoryPool* pool);
+
+}  // namespace ree_util
+}  // namespace arrow
+
+// Include internal utilities needed by template implementations below
+#include "arrow/util/ree_util_internal.h"
+
+namespace arrow {
+namespace ree_util {
+
+// Publish PhysicalIndexFinder from internal namespace
 template <typename RunEndCType>
 using PhysicalIndexFinder = internal::PhysicalIndexFinder<RunEndCType>;
